Clean up some comments
